Rename package.metadata.generate-rpm.repository to url and add fall back parameter package.metadata.generate-rpm.homepage to it.
The RPM variable "URL" is intended to store the URL of the package home page, not the repository URL. Therefore, the metadata key name is changed to "url" and set the fallback to "package.homepage" when the package key does not present. package.repository is also used for second fallback when neither package.metadata.generate-rpm.url nor package.homepage exists. Because today there is many projects where package.repository present but package.url not present.