give up to check the content of Config::table_to_dependencies().
w.r.t. #62, the fix is introduced in #63, which may pass in some environment but may not pass in some environment. Fundamentally, Dependency is bad because it is not derived from Eq but uses unsafe code for comparison. Also, the source code is simple and unlikely to have bugs, and there is no need to change it in the future. Therefore, we give up on unit test to check the problem areas.