Add --profile command line option
When expanding a source asset path, the "target/release/" prefix is now
remapped using profile in addition to target and target-dir. This enables
use of both alternative target-triple and build profiles.

For example, given the following asset configuration in Cargo.toml:

    [package.metadata.generate-rpm]
    assets = [
      { source = "target/release/my-exe", dest = "/usr/bin/", mode = "755" },
    ]

rpms can be generated for multiple target triples, such as:

    cargo generate-rpm --profile my-profile --target x86_64-unknown-linux-musl
    cargo generate-rpm --profile my-profile --target i668-unknown-linux-musl

with the paths for the "my-exe" executable expanded correctly to:

    target/x86_64-unknown-linux-musl/my-profile/my-exe
    target/i686-unknown-linux-musl/my-profile/my-exe

This behavior is consistent with cargo-deb, which also has a --profile
option for the same purpose.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>